requirements.txt
spotipy
flask
pandas
numpy
plotly
flask_login
flask_bcrypt
flask_sqlalchemy
spotipy
flask
pandas
matplotlib
flask-migrate
flask-login